Code formatted using Google's code guidelines for Android, App crash fixed, request access fixed and minor performance improvements.
