+++ .idea/ktfmt.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="KtfmtSettings">
+    <option name="enabled" value="true" />
+    <option name="uiFormatterStyle" value="Google (internal)" />
+  </component>
+</project>
