+++ app/src/main/AndroidManifest.xml
@@ -14,16 +14,16 @@
         android:theme="@style/Theme.Test"
         tools:targetApi="31" >
         <activity
-            android:name=".Implicit_Activity"
+            android:name=".ImplicitActivity"
             android:exported="false" />
         <activity
-            android:name=".Explicit_Activity"
+            android:name=".ExplicitActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".splashScreen"
+            android:name=".SplashScreen"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
