+++ app/src/main/java/com/example/test/ExplicitActivity.kt
@@ -1,0 +1,11 @@
+package com.example.test
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class ExplicitActivity : AppCompatActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_explicit)
+  }
+}
