--- app/src/main/java/com/example/test/Explicit_Activity.kt
@@ -1,11 +1,0 @@
-package com.example.test
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class Explicit_Activity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_explicit)
-    }
-}
