+++ app/src/main/java/com/example/test/SplashScreen.kt
@@ -1,0 +1,23 @@
+package com.example.test
+
+import android.content.Intent
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import androidx.appcompat.app.AppCompatActivity
+
+class SplashScreen : AppCompatActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_splash_screen)
+    Handler(Looper.getMainLooper())
+      .postDelayed(
+        {
+          intent = Intent(this, MainActivity::class.java)
+          startActivity(intent)
+          finish()
+        },
+        3000
+      )
+  }
+}
