--- app/src/main/java/com/example/test/splashScreen.kt
@@ -1,19 +1,0 @@
-package com.example.test
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-
-class splashScreen : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash_screen)
-        Handler(Looper.getMainLooper()).postDelayed({
-            intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-            finish()
-        },3000)
-    }
-}
