+++ app/src/main/res/layout/activity_explicit.xml
@@ -4,22 +4,19 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Explicit_Activity"
-    android:background="#2B2727">
+    android:background="#2B2727"
+    tools:context=".ExplicitActivity">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-
-        android:text="It is new Screen"
-        android:textColor="@color/white"
-        android:textSize="20dp"
         android:fontFamily="casual"
-
-        app:layout_constraintTop_toTopOf="parent"
+        android:text="@string/it_is_new_screen"
+        android:textColor="@color/white"
+        android:textSize="20sp"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        />
+        app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
